+++ CHANGELOG.md
@@ -6,7 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## [Unreleased]
 
-Nothing yet.
+### Change
+UI
+- The type of the parameter `onPointerEvent` has been changed to `((event: PointerEvent, charIndex: Int, tag: String?) -> Unit)?` (`charIndex: Int` is added).
 
 
 ## UI [2.2.0], DS [2.1.0] - 2025-04-16